refactor: date range selector
