+++ frontend/app/_home/home.tsx
@@ -2,36 +2,17 @@
 
 import { useEffect, useMemo, useState } from "react";
 import { useRouter } from "next/navigation";
-import { ChevronDown } from "lucide-react";
 
 import ScrollToTopButton from "@/components/navigation/scroll-to-top-button";
 import ArticleLoading from "@/components/news/article-loading";
 import NewsArticle from "@/components/news/news-article";
-import { useUpdateTopEventsPeriod } from "@/queries/user";
 import { useUserStore } from "@/store/user/user-store-provider";
 import { parseDate, toQueryDate } from "@/utils/date";
 import usePagination from "@/hooks/use-pagination";
 import Pagination from "@/components/navigation/pagination";
 import { getHomeEvents } from "@/queries/event";
 import { useQuery } from "@tanstack/react-query";
-
-const enum Period {
-  Day = 1,
-  Week = 7,
-  Month = 30,
-}
-
-const getDisplayValueFor = (period: Period) => {
-  switch (period) {
-    case Period.Day:
-      return "past day";
-    case Period.Month:
-      return "month";
-    case Period.Week:
-    default:
-      return "week";
-  }
-};
+import DateRangeSelector, { Period } from "@/app/_home/date-range-selector";
 
 const DEFAULT_EVENT_PERIOD = Period.Week;
 
@@ -40,7 +21,7 @@ const Home = () => {
   const [totalEventCount, setTotalEventCount] = useState<number | undefined>(
     undefined,
   );
-  const [showDropdown, setShowDropdown] = useState<boolean>(false);
+
   const [selectedPeriod, setSelectedPeriod] =
     useState<Period>(DEFAULT_EVENT_PERIOD);
 
@@ -48,7 +29,7 @@ const Home = () => {
     totalCount: totalEventCount,
   });
   const user = useUserStore((state) => state.user);
-  const updateTopEventsMutation = useUpdateTopEventsPeriod();
+
   const router = useRouter();
 
   const eventPeriod = useMemo(
@@ -82,17 +63,6 @@ const Home = () => {
     [events?.total_count],
   );
 
-  // Handle the option selection and close dropdown
-  const handleSelection = (period: Period) => {
-    if (period != selectedPeriod) {
-      // Update the text
-      setSelectedPeriod(period);
-      updateTopEventsMutation.mutate({ timePeriod: period });
-    }
-    // Close dropdown
-    setShowDropdown(false);
-  };
-
   if (!user!.categories.length) {
     router.push("/onboarding");
   }
@@ -113,41 +83,10 @@ const Home = () => {
               <span className="text-4xl 2xl:text-4xl font-bold text-primary-800">
                 What happened this&nbsp;
               </span>
-              <div className="relative">
-                <button
-                  className="flex items-center space-x-2 text-3xl 2xl:text-4xl font-bold hover:underline"
-                  onClick={() => setShowDropdown(!showDropdown)}
-                >
-                  <span className="text-4xl 2xl:text-4xl font-bold text-primary-800">
-                    {getDisplayValueFor(selectedPeriod)}
-                  </span>
-                  <ChevronDown className="w-4 h-4" />
-                </button>
-                {showDropdown && (
-                  <div className="absolute mt-2 w-40 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5">
-                    <div className="py-1">
-                      <button
-                        className="block w-full text-left px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-                        onClick={() => handleSelection(Period.Day)}
-                      >
-                        past day
-                      </button>
-                      <button
-                        className="block w-full text-left px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-                        onClick={() => handleSelection(Period.Week)}
-                      >
-                        week
-                      </button>
-                      <button
-                        className="block w-full text-left px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-                        onClick={() => handleSelection(Period.Month)}
-                      >
-                        month
-                      </button>
-                    </div>
-                  </div>
-                )}
-              </div>
+              <DateRangeSelector
+                selectedPeriod={selectedPeriod}
+                setSelectedPeriod={setSelectedPeriod}
+              />
             </div>
             <span className="text-primary text-lg">
               {parseDate(eventStartDate)} - {parseDate(new Date())}
